requirements.txt
pbr>=0.5.21,<1.0
M2Crypto
eventlet>=0.13.0
greenlet>=0.3.2
httplib2
iso8601>=0.1.8
kombu>=2.4.8
argparse
lxml>=2.3
netaddr
six>=1.4.1
sqlalchemy-migrate>=0.7.2,!=0.9.2
python-novaclient>=2.15.0
PasteDeploy>=1.5.0
Routes>=1.12.3,!=2.0
SQLAlchemy>=0.7.8,<=0.7.99
WebOb>=1.2.3,<1.3
python-keystoneclient>=0.3.2
python-swiftclient>=1.5
python-neutronclient>=2.3.4,<3
python-ceilometerclient>=1.0.5
python-cinderclient>=1.0.6
PyYAML>=3.1.0
paramiko>=1.8.0
Babel>=1.3
oslo.config>=1.2.0
